+++ src/cli.ts
@@ -19,14 +19,6 @@ export async function cli(args: string[]) {
   options = OptionsHelper.parseArguments(options, args);
   options = await OptionsHelper.promptForMissingArgs(options);
 
-  // console.log("Testing", {
-  //   ...options,
-  //   password: options.password ? "******" : undefined,
-  //   certificateBase64Encoded: options.certificateBase64Encoded
-  //     ? "******"
-  //     : undefined,
-  // })
-
   try {
     if (options.task === "help") {
       console.log("");